+++ lib/model/mock_post.dart
@@ -1,0 +1,18 @@
+import 'package:socially_app_flutter_ui/model/post_model.dart';
+
+final List<Post> mockPosts = [
+  Post(
+      post_id: 1,
+      user_id: 1,
+      content: 'This is a beautiful day!',
+      img_url: 'assets/images/building-1.jpg',
+      video_url: 'assets/images/building-1.jpg',
+      created_at: null),
+  Post(
+      post_id: 1,
+      user_id: 1,
+      content: 'This is a beautiful day!',
+      img_url: 'assets/images/building-1.jpg',
+      video_url: 'assets/images/building-1.jpg',
+      created_at: null),
+];
